Move SGX functionalities to separate file (from activity)
